PLT-3313: Extra module for `SemiArbitrary`
